requirements.txt
# Core Dependencies
torch==2.6.0+cu118
torchvision==0.17.0+cu118
transformers==4.50.3
Pillow==9.3.0
PyYAML==6.0.1
pandas==2.0.3
numpy==1.24.1
jupyter==1.0.0
matplotlib==3.7.3
seaborn==0.12.2
opencv-python==4.9.0.80  # For image processing
joblib==1.3.2  # For caching support
psutil==5.9.8  # For memory monitoring in tests

# Configuration and Type Checking
pydantic==2.6.1  # For data validation and settings management
typing-extensions==4.9.0
mypy==1.8.0
types-PyYAML==6.0.12.12
types-Pillow==10.2.0.20240311  # Type stubs for Pillow
types-pandas==2.2.0.20240311  # Type stubs for pandas
injector==0.21.0  # For dependency injection support

# Data Processing
scikit-learn==1.3.2
scipy==1.15.2
tqdm==4.67.1
python-magic==0.4.27
python-magic-bin==0.4.14

# Testing and Development
pytest==8.3.5
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5
pytest-dependency==0.5.1  # For testing dependency injection
pytest-factoryboy==2.7.0  # For factory pattern testing
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
nbsphinx==0.9.3
myst-parser==2.0.0

# Model Dependencies
# Uncomment as needed for specific model implementations
# doctr==2.4.0
# llama-cpp-python==0.2.23
# pixtral==0.1.0

# Note: System Requirements
# - CUDA 11.8 is required for PyTorch
# - GPU: NVIDIA H100 NVL recommended
# - Minimum GPU memory: 93.00 GB
# - Python version: 3.13.2 or higher

# Development Environment
# - VS Code with Python extension recommended
# - Install pre-commit hooks for automatic formatting
# - Use virtual environment (venv) for isolation
# - Dependency injection pattern used throughout codebase
